+++ package.json
@@ -6,6 +6,8 @@
     "node-sass": "^4.13.1",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "2.1.1"
   },
   "scripts": {
