+++ src/views/ArticlesView/ArticleView.js
@@ -1,0 +1,9 @@
+import React, { Component } from "react";
+
+class ArticleView extends Component {
+  render() {
+    return <p>This is ArticleView page</p>;
+  }
+}
+
+export default ArticleView;
