+++ src/views/NotesView/NotesView.js
@@ -1,0 +1,9 @@
+import React, { Component } from "react";
+
+class NotesView extends Component {
+  render() {
+    return <p>This is NotesView page</p>;
+  }
+}
+
+export default NotesView;
